fixed requested changes
