+++ apps/web/components/shared/kyc-4/file-upload-area.tsx
@@ -1,29 +1,31 @@
-import { Upload } from 'lucide-react'
-import { useRef } from 'react'
-import type { useToast } from '~/components/base/toast'
-import { cn } from '~/lib/utils'
-import type { DocumentType } from '~/components/types'
+import { Upload } from 'lucide-react';
+import { useRef } from 'react';
+import type { useToast } from '~/components/base/toast';
+import { cn } from '~/lib/utils';
+import type { DocumentType } from '~/types';
+
+type documentType = DocumentType;
 
 interface FileUploadAreaProps {
-  isProcessing: boolean
-  documentType: DocumentType
+  isProcessing: boolean;
+  documentType: documentType;
   handleFileSelect: (
     e: React.ChangeEvent<HTMLInputElement>,
-    documentType: DocumentType,
+    documentType: string,
     handleFileUploadBound: (file: File) => void,
     setFile: React.Dispatch<React.SetStateAction<File | null>>,
-    toast: ReturnType<typeof useToast>['toast']
-  ) => void
-  handleFileUploadBound: (file: File) => void
-  setFile: React.Dispatch<React.SetStateAction<File | null>>
-  toast: ReturnType<typeof useToast>['toast']
+    toast: ReturnType<typeof useToast>['toast'],
+  ) => void;
+  handleFileUploadBound: (file: File) => void;
+  setFile: React.Dispatch<React.SetStateAction<File | null>>;
+  toast: ReturnType<typeof useToast>['toast'];
   handleDrop: (
-    e: React.DragEvent<HTMLElement>,  // Changed to HTMLElement to be more generic
-    documentType: DocumentType,
+    e: React.DragEvent<HTMLDivElement>,
+    documentType: documentType,
     handleFileUploadBound: (file: File) => void,
     setFile: React.Dispatch<React.SetStateAction<File | null>>,
-    toast: ReturnType<typeof useToast>['toast']
-  ) => void
+    toast: ReturnType<typeof useToast>['toast'],
+  ) => void;
 }
 
 export const FileUploadArea: React.FC<FileUploadAreaProps> = ({
@@ -35,31 +37,33 @@ export const FileUploadArea: React.FC<FileUploadAreaProps> = ({
   setFile,
   toast,
 }) => {
-  const fileUploadRef = useRef<HTMLInputElement | null>(null)
-  
-  const handleClick = () => {
-    if (isProcessing || !fileUploadRef.current) return
-    fileUploadRef.current.click()
-  }
-  
+  const fileUploadRef = useRef<HTMLInputElement | null>(null);
+
   return (
-    <button
-      type="button"
+    <div
       className={cn(
-        'border-2 border-dashed rounded-lg p-8 text-center w-full',
+        'border-2 border-dashed rounded-lg p-8 text-center',
         'transition-all duration-300 ease-in-out transform',
         'hover:scale-[1.02] hover:border-black hover:bg-gray-50',
         'hover:shadow-lg cursor-pointer relative',
-        { 'opacity-50 pointer-events-none': isProcessing }
+        { 'opacity-50 pointer-events-none': isProcessing },
       )}
-      onDrop={(e) => {
-        e.preventDefault()
+      onDrop={(e) =>
         handleDrop(e, documentType, handleFileUploadBound, setFile, toast)
-      }}
+      }
       onDragOver={(e) => e.preventDefault()}
-      onClick={handleClick}
-      disabled={isProcessing}
-      aria-label="File upload area"
+      onClick={() => {
+        if (isProcessing || !fileUploadRef.current) return;
+        fileUploadRef.current.click();
+      }}
+      onKeyUp={(e) => {
+        if (isProcessing || !fileUploadRef.current) return;
+        if (e.key === 'Enter' || e.key === ' ') {
+          fileUploadRef.current.click();
+        }
+      }}
+      tabIndex={0}
+      role="button"
     >
       <input
         ref={fileUploadRef}
@@ -74,7 +78,7 @@ export const FileUploadArea: React.FC<FileUploadAreaProps> = ({
             documentType,
             handleFileUploadBound,
             setFile,
-            toast
+            toast,
           )
         }
         disabled={isProcessing}
@@ -82,6 +86,6 @@ export const FileUploadArea: React.FC<FileUploadAreaProps> = ({
       <Upload className="mx-auto mb-4 h-12 w-12 text-gray-400" />
       <h3 className="text-lg font-medium">Click to upload or drag and drop</h3>
       <p className="text-sm text-gray-500">JPG or PNG images only</p>
-    </button>
-  )
-}
+    </div>
+  );
+};
