+++ apps/web/components/shared/kyc-4/ocr-processor.tsx
@@ -1,0 +1,104 @@
+import { ArrowLeft, ArrowRight } from 'lucide-react'
+import type React from 'react'
+import { Button } from '~/components/base/button'
+import { Toast, type useToast } from '~/components/base/toast'
+import type { DocumentType, ExtractedData } from '~/types'
+
+interface OCRProcessorProps {
+	isProcessing: boolean
+	onBack?: () => void
+	onNext?: (data: {
+		documentType: DocumentType
+		extractedData: ExtractedData
+	}) => void
+	extractedData: ExtractedData | null
+	documentType: DocumentType
+	handleContinue: (
+		extractedData: ExtractedData | null,
+		documentType: DocumentType,
+		onNext: (data: {
+			documentType: DocumentType
+			extractedData: ExtractedData
+		}) => void,
+		validateDocument: (data: ExtractedData) => {
+			isValid: boolean
+			errors: string[]
+		},
+		toast: ReturnType<typeof useToast>['toast'],
+		setValidationErrors: React.Dispatch<React.SetStateAction<string[]>>,
+	) => void
+	validateDocument: (data: ExtractedData) => {
+		isValid: boolean
+		errors: string[]
+	}
+	toast: ReturnType<typeof useToast>['toast']
+	setValidationErrors: React.Dispatch<React.SetStateAction<string[]>>
+}
+
+export const OCRProcessor: React.FC<OCRProcessorProps> = ({
+	isProcessing,
+	onBack,
+	onNext,
+	extractedData,
+	documentType,
+	validateDocument,
+	toast,
+	setValidationErrors,
+}) => {
+	const handleContinue = (
+		extractedData: ExtractedData | null,
+		documentType: DocumentType,
+		onNext: (data: {
+			documentType: DocumentType
+			extractedData: ExtractedData
+		}) => void,
+		validateDocument: (data: ExtractedData) => {
+			isValid: boolean
+			errors: string[]
+		},
+		toast: ReturnType<typeof useToast>['toast'],
+		setValidationErrors: React.Dispatch<React.SetStateAction<string[]>>,
+	) => {
+		if (!extractedData) {
+			toast({ title: 'Error', description: 'No data extracted' })
+			return
+		}
+
+		const validation = validateDocument(extractedData)
+		if (!validation.isValid) {
+			setValidationErrors(validation.errors)
+			toast({
+				title: 'Validation Error',
+				description: 'Document validation failed',
+			})
+			return
+		}
+
+		onNext({ documentType, extractedData })
+	}
+
+	return (
+		<div className="flex justify-end space-x-4">
+			<Button variant="outline" onClick={onBack} disabled={isProcessing}>
+				<ArrowLeft className="mr-2 h-4 w-4" /> Back
+			</Button>
+			<Button
+				onClick={() => {
+					if (onNext) {
+						handleContinue(
+							extractedData,
+							documentType,
+							onNext,
+							validateDocument,
+							toast,
+							setValidationErrors,
+						)
+					}
+				}}
+				disabled={isProcessing || !extractedData || !documentType}
+			>
+				Continue <ArrowRight className="ml-2 h-4 w-4" />
+			</Button>
+		</div>
+	)
+}
