--- apps/web/components/shared/kyc/kyc-4/OCR-processor.tsx
@@ -1,73 +1,0 @@
-import { ArrowLeft, ArrowRight } from 'lucide-react'
-import type React from 'react'
-import { Button } from '~/components/base/button'
-import type { useToast } from '~/components/base/toast'
-import type { DocumentType, ExtractedData } from '~/components/types'
-
-// Define function types without parameter names to avoid unused variable warnings
-type ValidateDocumentFn = (_: ExtractedData) => {
-  isValid: boolean
-  errors: string[]
-}
-
-type OnNextFn = (_: {
-  documentType: DocumentType
-  extractedData: ExtractedData
-}) => void
-
-interface OCRProcessorProps {
-  isProcessing: boolean
-  onBack?: () => void
-  onNext?: OnNextFn
-  extractedData: ExtractedData | null
-  documentType: DocumentType
-  validateDocument: ValidateDocumentFn
-  toast: ReturnType<typeof useToast>['toast']
-  setValidationErrors: React.Dispatch<React.SetStateAction<string[]>>
-}
-
-export const OCRProcessor: React.FC<OCRProcessorProps> = ({
-  isProcessing,
-  onBack,
-  onNext,
-  extractedData,
-  documentType,
-  validateDocument,
-  toast,
-  setValidationErrors,
-}) => {
-  const handleContinueClick = () => {
-    if (!extractedData) {
-      toast({ title: 'Error', description: 'No data extracted' })
-      return
-    }
-    
-    const validation = validateDocument(extractedData)
-    if (!validation.isValid) {
-      setValidationErrors(validation.errors)
-      toast({
-        title: 'Validation Error',
-        description: 'Document validation failed',
-      })
-      return
-    }
-    
-    if (onNext) {
-      onNext({ documentType, extractedData })
-    }
-  }
-
-  return (
-    <div className="flex justify-end space-x-4">
-      <Button variant="outline" onClick={onBack} disabled={isProcessing}>
-        <ArrowLeft className="mr-2 h-4 w-4" /> Back
-      </Button>
-      <Button
-        onClick={handleContinueClick}
-        disabled={isProcessing || !extractedData || !documentType}
-      >
-        Continue <ArrowRight className="ml-2 h-4 w-4" />
-      </Button>
-    </div>
-  )
-}
