+++ apps/web/hooks/kyc/use-document-validation.ts
@@ -1,6 +1,9 @@
 'use client'
 import { useCallback, useState } from 'react'
-import type { DocumentType, ExtractedData } from '../../components/shared/kyc/kyc-2/types'
+import type {
+	DocumentType,
+	ExtractedData,
+} from '../../components/shared/kyc/kyc-2/types'
 import { DocumentPatterns } from '../../components/shared/kyc/kyc-2/types'
 
 export function useDocumentValidation(documentType: DocumentType) {
