--- apps/web/lib/types/kyc-4.types.ts
@@ -1,7 +1,0 @@
-export type DocumentType = 'utility' | 'bank' | 'government' | ''
-
-export interface ExtractedData {
-	text: string
-	date: string | null
-	address: string | null
-}
