--- apps/web/lib/utils/index.ts
@@ -1,3 +1,0 @@
-export * from './doc-utils/upload-handler'
-export * from './doc-utils/validation'
-export * from './doc-utils/extraction'
