+++ packages/lib/package.json
@@ -1,0 +1,11 @@
+{
+	"name": "@packages/lib",
+	"version": "0.0.1",
+	"private": true,
+	"type": "module",
+	"main": "./src/index.ts",
+	"types": "./src/index.ts",
+	"license": "MIT",
+	"dependencies": {},
+	"devDependencies": {}
+}
